+++ bin/citrea/Makefile
@@ -1,0 +1,5 @@
+mock-e2e:
+	cargo test mock_e2e
+
+test-bitcoin:
+	cargo test bitcoin_e2e
