+++ bin/citrea/tests/bitcoin/bitcoin_test.rs
