+++ bin/citrea/tests/bitcoin/full_node.rs
