+++ bin/citrea/tests/bitcoin/light_client_test.rs
