+++ bin/citrea/tests/bitcoin/sequencer_commitments.rs
