+++ bin/citrea/tests/bitcoin/test-data/gen-proof-input-genesis/evm.json
