+++ bin/citrea/tests/bitcoin/test-data/gen-proof-input-genesis/soft_confirmation_rule_enforcer.json
