+++ bin/citrea/tests/bitcoin_e2e.rs
@@ -1,0 +1,2 @@
+mod bitcoin;
+mod common;
