+++ bin/citrea/tests/common/constants.rs
@@ -1,0 +1,3 @@
+pub const TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT: u64 = 1000;
+pub const DEFAULT_PROOF_WAIT_DURATION: u64 = 600; // 10 minutes
+pub const TEST_DATA_GENESIS_PATH: &str = "../../resources/test-data/integration-tests";
