+++ bin/citrea/tests/e2e/sequencer_behaviour.rs
@@ -13,14 +13,14 @@ use reth_primitives::BlockNumberOrTag;
 use sov_mock_da::{MockAddress, MockDaService, MockDaSpec};
 use tokio::time::sleep;
 
-use crate::e2e::{initialize_test, TestConfig};
-use crate::evm::{init_test_rollup, make_test_client};
-use crate::test_client::TestClient;
-use crate::test_helpers::{
+use crate::common::client::TestClient;
+use crate::common::helpers::{
     create_default_rollup_config, start_rollup, tempdir_with_children, wait_for_commitment,
     wait_for_l1_block, wait_for_l2_block, NodeMode,
 };
-use crate::TEST_DATA_GENESIS_PATH;
+use crate::common::{make_test_client, TEST_DATA_GENESIS_PATH};
+use crate::e2e::{initialize_test, TestConfig};
+use crate::evm::init_test_rollup;
 
 /// Run the sequencer.
 /// Create some blocks.
