+++ bin/citrea/tests/e2e/system_transactions.rs
@@ -8,8 +8,8 @@ use reth_primitives::BlockNumberOrTag;
 use sov_mock_da::{MockAddress, MockDaService};
 use sov_rollup_interface::services::da::DaService;
 
+use crate::common::helpers::{tempdir_with_children, wait_for_l1_block, wait_for_l2_block};
 use crate::e2e::{initialize_test, TestConfig};
-use crate::test_helpers::{tempdir_with_children, wait_for_l1_block, wait_for_l2_block};
 
 /// Trigger system transactions.
 /// Ask the sequencer and the full node for blocks.
