+++ bin/citrea/tests/evm/archival_state.rs
@@ -8,12 +8,12 @@ use citrea_stf::genesis_config::GenesisPaths;
 use reth_primitives::{BlockId, BlockNumberOrTag};
 use tokio::time::sleep;
 
-use crate::evm::init_test_rollup;
-use crate::test_client::TestClient;
-use crate::test_helpers::{
+use crate::common::client::TestClient;
+use crate::common::helpers::{
     create_default_rollup_config, start_rollup, tempdir_with_children, wait_for_l2_block, NodeMode,
 };
-use crate::TEST_DATA_GENESIS_PATH;
+use crate::common::TEST_DATA_GENESIS_PATH;
+use crate::evm::init_test_rollup;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_archival_state() -> Result<(), anyhow::Error> {
