+++ bin/citrea/tests/evm/fee.rs
@@ -4,11 +4,11 @@ use citrea_common::SequencerConfig;
 use citrea_stf::genesis_config::GenesisPaths;
 use reth_primitives::BlockNumberOrTag;
 
-use crate::evm::init_test_rollup;
-use crate::test_helpers::{
+use crate::common::helpers::{
     create_default_rollup_config, start_rollup, tempdir_with_children, wait_for_l2_block, NodeMode,
 };
-use crate::TEST_DATA_GENESIS_PATH;
+use crate::common::TEST_DATA_GENESIS_PATH;
+use crate::evm::init_test_rollup;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_minimum_base_fee() -> Result<(), anyhow::Error> {
