+++ bin/citrea/tests/evm/subscription.rs
@@ -12,13 +12,13 @@ use citrea_evm::{Filter, LogResponse};
 use citrea_stf::genesis_config::GenesisPaths;
 use tokio::time::sleep;
 
-use crate::evm::make_test_client;
-use crate::test_client::TestClient;
-use crate::test_helpers::{
+use crate::common::client::TestClient;
+use crate::common::helpers::{
     create_default_rollup_config, start_rollup, tempdir_with_children, wait_for_l2_block, NodeMode,
 };
-use crate::{
-    TEST_DATA_GENESIS_PATH, TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
+use crate::common::{
+    make_test_client, TEST_DATA_GENESIS_PATH,
+    TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
 };
 
 #[tokio::test(flavor = "multi_thread")]
