+++ bin/citrea/tests/soft_confirmation_rule_enforcer/mod.rs
@@ -5,13 +5,14 @@ use citrea_stf::genesis_config::GenesisPaths;
 use sov_mock_da::{MockAddress, MockDaService};
 use tokio::time::sleep;
 
-use crate::evm::make_test_client;
 // use citrea::initialize_logging;
-use crate::test_helpers::{
+use crate::common::helpers::{
     create_default_rollup_config, start_rollup, tempdir_with_children, wait_for_l1_block,
     wait_for_l2_block, NodeMode,
 };
-use crate::TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT;
+use crate::common::{
+    make_test_client, TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
+};
 
 /// Transaction with equal nonce to last tx should not be accepted by mempool.
 #[tokio::test(flavor = "multi_thread")]
